split_fileset moved to dataset_tools
